+++ NEWS.md
@@ -1,3 +1,8 @@
+# rxode2 development version
+
+- A bug was fixed so that the `zeroRe()` function works with correlated omega
+  values.
+
 # rxode2 2.0.12
 
 ## New features